Feature: Sending email

+++ app/src/main/java/malaria/com/malaria/models/Result.java
@@ -8,7 +8,7 @@ public class Result {
     /**
      *
      */
-    private double parasites_per_microlitre;
+    private int parasites_per_microlitre;
 
     /**
      * Default constructor
@@ -16,11 +16,11 @@ public class Result {
     public Result() {
     }
 
-    public double getParasites_per_microlitre() {
+    public int getParasites_per_microlitre() {
         return parasites_per_microlitre;
     }
 
-    public void setParasites_per_microlitre(double parasites_per_microlitre) {
+    public void setParasites_per_microlitre(int parasites_per_microlitre) {
         this.parasites_per_microlitre = parasites_per_microlitre;
     }
 }
